pyproject.toml
[project]
name = "simpleagentshealthinsurance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv==1.2.2",
    "simple-agents-py==0.3.9",
    "streamlit==1.56.0",
]
